refactor(core): register URI protocol once, remove from WebviewAttributes
