+++ .changes/refactor-register-uri-scheme-protocol.md
@@ -1,0 +1,6 @@
+---
+"tauri": patch
+"tauri-runtime": patch
+---
+
+**Breaking change:** Removed `register_uri_scheme_protocol` from the `WebviewAttibutes` struct.
