+++ core/tauri-runtime-wry/src/lib.rs
@@ -2133,6 +2133,7 @@ fn create_webview(
   #[allow(unused_mut)]
   let PendingWindow {
     webview_attributes,
+    uri_scheme_protocols,
     mut window_builder,
     rpc_handler,
     file_drop_handler,
@@ -2183,7 +2184,7 @@ fn create_webview(
       handler,
     ));
   }
-  for (scheme, protocol) in webview_attributes.uri_scheme_protocols {
+  for (scheme, protocol) in uri_scheme_protocols {
     webview_builder = webview_builder.with_custom_protocol(scheme, move |url| {
       protocol(url)
         .map(|data| {
